api/requirements.txt
# ============================================
# 儿童表现记录系统 - 后端API依赖
# Python 3.10+
# ============================================

# Web 框架
fastapi==0.115.6
uvicorn[standard]==0.34.0

# 数据库
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# 数据验证
pydantic==2.10.3
pydantic-settings==2.7.0

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
bcrypt==4.2.1

# HTTP 客户端 (微信 API 调用)
httpx==0.28.1

# 工具
python-dotenv==1.0.1
python-dateutil==2.9.0

# CORS
starlette==0.41.3

# 日志
loguru==0.7.3